requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers